requirements.txt
allure-pytest
jsonpath
pytest
pytest-html
requests
pyyaml